Change params to JSONObject and implement ResultCallable
Changed AsyncUploader to send up all params.
Changed AsyncUploader to implement ResultCallable, not sure the
functions should do.
Alternatively, could make ResultCallable a class rather than an
interface as it was initially designed.
