+++ app/src/main/java/cs5150athletetracking/com/athletetracking/ResultCallable.java
@@ -1,0 +1,10 @@
+package cs5150athletetracking.com.athletetracking;
+
+/**
+ * Created by mag94 on 10/20/15.
+ */
+public interface ResultCallable {
+    public void success();
+
+    public void failure();
+}
